# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: VetOS API
  version: 0.1.0
  description: |
    Cookie-authenticated clinic API. Implemented paths include auth, clinic profile,
    user administration, branch list, and health. Paths marked x-status planned are
    the remaining first vertical slice contract.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Cookies set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token
      security: []
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "200":
          description: Cookies rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      parameters:
        - $ref: "#/components/parameters/CsrfHeader"
      responses:
        "204":
          description: Cookies cleared
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Current user
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionUser"
        "401":
          $ref: "#/components/responses/Error"
...
components:
  parameters:
    CsrfHeader:
      in: header
      name: X-XSRF-TOKEN
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: vetos_access
  schemas:
    ApiError:
      type: object
      required: [code, message, details, timestamp, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    SessionUser:
      type: object
      required: [id, clinicId, displayName, roles]
      properties:
        id:
          type: string
          format: uuid
        clinicId:
          type: string
          format: uuid
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
...
security:
  - cookieAuth: []